Move SharedTreeImpl and ISharedTree
