+++ packages/dds/tree/src/test/snapshots/snapshotTestScenarios.ts
@@ -7,15 +7,10 @@ import { strict as assert } from "node:assert";
 import { MockFluidDataStoreRuntime } from "@fluidframework/test-runtime-utils/internal";
 
 import { typeboxValidator } from "../../external-utilities/index.js";
-import {
-	getBranch,
-	type ISharedTree,
-	type SharedTreeOptions,
-	Tree,
-} from "../../shared-tree/index.js";
+import { getBranch, type SharedTreeOptions, Tree } from "../../shared-tree/index.js";
 import { createSnapshotCompressor, TestTreeProviderLite } from "../utils.js";
 import { SchemaFactory, TreeViewConfiguration } from "../../simple-tree/index.js";
-import { configuredSharedTree } from "../../treeFactory.js";
+import { configuredSharedTree, type ISharedTree } from "../../treeFactory.js";
 import type { IChannel } from "@fluidframework/datastore-definitions/internal";
 
 const enableSchemaValidation = true;
